relocated camera in camera.rs
